Spectacle for visually impaired person
This is a spectacle built for partially or fully visually disabled
person. When any obstacle is detected in the ground or left or right on
the walking way of the user then the vibrators alert the person through
vibration to avoid that path to go. Blind person can move easily